+++ python/pyvelox/__init__.py
@@ -12,4 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_subdirectory(src)
+from .legacy import *
+from .legacy import __version__
