+++ python/src/arrow/arrow.cpp
@@ -23,8 +23,8 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include "python/init/PyInit.h"
-#include "python/vector/PyVector.h"
+#include "python/src/init/PyInit.h"
+#include "python/src/vector/PyVector.h"
 #include "velox/vector/arrow/Bridge.h"
 
 namespace py = pybind11;
