+++ python/src/file/file.cpp
@@ -17,8 +17,8 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include "python/file/PyFile.h"
-#include "python/lib/PyInit.h"
+#include "python/src/file/PyFile.h"
+#include "python/src/lib/PyInit.h"
 
 namespace py = pybind11;
