+++ python/src/legacy/serde.h
