+++ python/src/legacy/signatures.cpp
