+++ python/src/runner/PyLocalRunner.cpp
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-#include "python/runner/PyLocalRunner.h"
+#include "python/src/runner/PyLocalRunner.h"
 
 #include <pybind11/stl.h>
-#include "python/vector/PyVector.h"
+#include "python/src/vector/PyVector.h"
 #include "velox/connectors/hive/HiveConnectorSplit.h"
 #include "velox/core/PlanNode.h"
 #include "velox/dwio/common/Options.h"
