+++ python/src/vector/vector.pyi
