+++ python/test/test_expression.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import pyvelox.legacy as pv
+import pyvelox as pv
 import unittest
