+++ scripts/signature.py
@@ -20,7 +20,7 @@ from typing import Any
 
 from deepdiff import DeepDiff
 
-import pyvelox.legacy as pv
+import pyvelox as pv
 
 # Utility to export and diff function signatures.
